Enum rusoto_codedeploy::CreateDeploymentError

pub enum CreateDeploymentError {
    ApplicationDoesNotExist(String),
    ApplicationNameRequired(String),
    DeploymentConfigDoesNotExist(String),
    DeploymentGroupDoesNotExist(String),
    DeploymentGroupNameRequired(String),
    DeploymentLimitExceeded(String),
    DescriptionTooLong(String),
    InvalidApplicationName(String),
    InvalidAutoRollbackConfig(String),
    InvalidAutoScalingGroup(String),
    InvalidDeploymentConfigName(String),
    InvalidDeploymentGroupName(String),
    InvalidFileExistsBehavior(String),
    InvalidGitHubAccountToken(String),
    InvalidIgnoreApplicationStopFailuresValue(String),
    InvalidLoadBalancerInfo(String),
    InvalidRevision(String),
    InvalidRole(String),
    InvalidTargetInstances(String),
    InvalidUpdateOutdatedInstancesOnlyValue(String),
    RevisionDoesNotExist(String),
    RevisionRequired(String),
    Throttling(String),
}

Errors returned by CreateDeployment

Variants
ApplicationDoesNotExist(String)

The application does not exist with the IAM user or AWS account.

ApplicationNameRequired(String)

The minimum number of required application names was not specified.

DeploymentConfigDoesNotExist(String)

The deployment configuration does not exist with the IAM user or AWS account.

DeploymentGroupDoesNotExist(String)

The named deployment group with the IAM user or AWS account does not exist.

DeploymentGroupNameRequired(String)

The deployment group name was not specified.

DeploymentLimitExceeded(String)

The number of allowed deployments was exceeded.

DescriptionTooLong(String)

The description is too long.

InvalidApplicationName(String)

The application name was specified in an invalid format.

InvalidAutoRollbackConfig(String)

The automatic rollback configuration was specified in an invalid format. For example, automatic rollback is enabled, but an invalid triggering event type or no event types were listed.

InvalidAutoScalingGroup(String)

The Auto Scaling group was specified in an invalid format or does not exist.

InvalidDeploymentConfigName(String)

The deployment configuration name was specified in an invalid format.

InvalidDeploymentGroupName(String)

The deployment group name was specified in an invalid format.

InvalidFileExistsBehavior(String)

An invalid fileExistsBehavior option was specified to determine how AWS CodeDeploy handles files or directories that already exist in a deployment target location, but weren't part of the previous successful deployment. Valid values include "DISALLOW," "OVERWRITE," and "RETAIN."

InvalidGitHubAccountToken(String)

The GitHub token is not valid.

InvalidIgnoreApplicationStopFailuresValue(String)

The IgnoreApplicationStopFailures value is invalid. For AWS Lambda deployments, false is expected. For EC2/On-premises deployments, true or false is expected.

InvalidLoadBalancerInfo(String)

An invalid load balancer name, or no load balancer name, was specified.

InvalidRevision(String)

The revision was specified in an invalid format.

InvalidRole(String)

The service role ARN was specified in an invalid format. Or, if an Auto Scaling group was specified, the specified service role does not grant the appropriate permissions to Amazon EC2 Auto Scaling.

InvalidTargetInstances(String)

The target instance configuration is invalid. Possible causes include:

Configuration data for target instances was entered for an in-place deployment.

The limit of 10 tags for a tag type was exceeded.

The combined length of the tag names exceeded the limit.

A specified tag is not currently applied to any instances.

InvalidUpdateOutdatedInstancesOnlyValue(String)

The UpdateOutdatedInstancesOnly value is invalid. For AWS Lambda deployments, false is expected. For EC2/On-premises deployments, true or false is expected.

RevisionDoesNotExist(String)

The named revision does not exist with the IAM user or AWS account.

RevisionRequired(String)

The revision ID was not specified.

Throttling(String)

An API function was called too frequently.

Methods

impl CreateDeploymentError

pub fn from_response(
res: BufferedHttpResponse
) -> RusotoError<CreateDeploymentError>