Enum rusoto_codedeploy::DeleteDeploymentGroupError

pub enum DeleteDeploymentGroupError {
    ApplicationNameRequired(String),
    DeploymentGroupNameRequired(String),
    InvalidApplicationName(String),
    InvalidDeploymentGroupName(String),
    InvalidRole(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by DeleteDeploymentGroup

Variants
ApplicationNameRequired(String)

The minimum number of required application names was not specified.

DeploymentGroupNameRequired(String)

The deployment group name was not specified.

InvalidApplicationName(String)

The application name was specified in an invalid format.

InvalidDeploymentGroupName(String)

The deployment group name was specified in an invalid format.

InvalidRole(String)

The service role ARN was specified in an invalid format. Or, if an Auto Scaling group was specified, the specified service role does not grant the appropriate permissions to Auto Scaling.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl DeleteDeploymentGroupError

fn from_body(body: &str) -> DeleteDeploymentGroupError