Enum rusoto_codedeploy::RegisterOnPremisesInstanceError

pub enum RegisterOnPremisesInstanceError {
    IamArnRequired(String),
    IamSessionArnAlreadyRegistered(String),
    IamUserArnAlreadyRegistered(String),
    IamUserArnRequired(String),
    InstanceNameAlreadyRegistered(String),
    InstanceNameRequired(String),
    InvalidIamSessionArn(String),
    InvalidIamUserArn(String),
    InvalidInstanceName(String),
    MultipleIamArnsProvided(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by RegisterOnPremisesInstance

Variants
IamArnRequired(String)

No IAM ARN was included in the request. You must use an IAM session ARN or IAM user ARN in the request.

IamSessionArnAlreadyRegistered(String)

The request included an IAM session ARN that has already been used to register a different instance.

IamUserArnAlreadyRegistered(String)

The specified IAM user ARN is already registered with an on-premises instance.

IamUserArnRequired(String)

An IAM user ARN was not specified.

InstanceNameAlreadyRegistered(String)

The specified on-premises instance name is already registered.

InstanceNameRequired(String)

An on-premises instance name was not specified.

InvalidIamSessionArn(String)

The IAM session ARN was specified in an invalid format.

InvalidIamUserArn(String)

The IAM user ARN was specified in an invalid format.

InvalidInstanceName(String)

The specified on-premises instance name was specified in an invalid format.

MultipleIamArnsProvided(String)

Both an IAM user ARN and an IAM session ARN were included in the request. Use only one ARN type.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl RegisterOnPremisesInstanceError

fn from_body(body: &str) -> RegisterOnPremisesInstanceError