Enum rusoto_codedeploy::UpdateDeploymentGroupError

pub enum UpdateDeploymentGroupError {
    AlarmsLimitExceeded(String),
    ApplicationDoesNotExist(String),
    ApplicationNameRequired(String),
    DeploymentConfigDoesNotExist(String),
    DeploymentGroupAlreadyExists(String),
    DeploymentGroupDoesNotExist(String),
    DeploymentGroupNameRequired(String),
    InvalidAlarmConfig(String),
    InvalidApplicationName(String),
    InvalidAutoRollbackConfig(String),
    InvalidAutoScalingGroup(String),
    InvalidBlueGreenDeploymentConfiguration(String),
    InvalidDeploymentConfigName(String),
    InvalidDeploymentGroupName(String),
    InvalidDeploymentStyle(String),
    InvalidEC2TagCombination(String),
    InvalidEC2Tag(String),
    InvalidInput(String),
    InvalidLoadBalancerInfo(String),
    InvalidOnPremisesTagCombination(String),
    InvalidRole(String),
    InvalidTag(String),
    InvalidTriggerConfig(String),
    LifecycleHookLimitExceeded(String),
    TagSetListLimitExceeded(String),
    TriggerTargetsLimitExceeded(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by UpdateDeploymentGroup

Variants
AlarmsLimitExceeded(String)

The maximum number of alarms for a deployment group (10) was exceeded.

ApplicationDoesNotExist(String)

The application does not exist with the applicable IAM user or AWS account.

ApplicationNameRequired(String)

The minimum number of required application names was not specified.

DeploymentConfigDoesNotExist(String)

The deployment configuration does not exist with the applicable IAM user or AWS account.

DeploymentGroupAlreadyExists(String)

A deployment group with the specified name already exists with the applicable IAM user or AWS account.

DeploymentGroupDoesNotExist(String)

The named deployment group does not exist with the applicable IAM user or AWS account.

DeploymentGroupNameRequired(String)

The deployment group name was not specified.

InvalidAlarmConfig(String)

The format of the alarm configuration is invalid. Possible causes include:

The alarm list is null.

The alarm object is null.

The alarm name is empty or null or exceeds the 255 character limit.

Two alarms with the same name have been specified.

The alarm configuration is enabled but the alarm list is empty.

InvalidApplicationName(String)

The application name was specified in an invalid format.

InvalidAutoRollbackConfig(String)

The automatic rollback configuration was specified in an invalid format. For example, automatic rollback is enabled but an invalid triggering event type or no event types were listed.

InvalidAutoScalingGroup(String)

The Auto Scaling group was specified in an invalid format or does not exist.

InvalidBlueGreenDeploymentConfiguration(String)

The configuration for the blue/green deployment group was provided in an invalid format. For information about deployment configuration format, see CreateDeploymentConfig.

InvalidDeploymentConfigName(String)

The deployment configuration name was specified in an invalid format.

InvalidDeploymentGroupName(String)

The deployment group name was specified in an invalid format.

InvalidDeploymentStyle(String)

An invalid deployment style was specified. Valid deployment types include "IN_PLACE" and "BLUE_GREEN". Valid deployment options include "WITH_TRAFFIC_CONTROL" and "WITHOUT_TRAFFIC_CONTROL".

InvalidEC2TagCombination(String)

A call was submitted that specified both Ec2TagFilters and Ec2TagSet, but only one of these data types can be used in a single call.

InvalidEC2Tag(String)

The tag was specified in an invalid format.

InvalidInput(String)

The specified input was specified in an invalid format.

InvalidLoadBalancerInfo(String)

An invalid load balancer name, or no load balancer name, was specified.

InvalidOnPremisesTagCombination(String)

A call was submitted that specified both OnPremisesTagFilters and OnPremisesTagSet, but only one of these data types can be used in a single call.

InvalidRole(String)

The service role ARN was specified in an invalid format. Or, if an Auto Scaling group was specified, the specified service role does not grant the appropriate permissions to Auto Scaling.

InvalidTag(String)

The specified tag was specified in an invalid format.

InvalidTriggerConfig(String)

The trigger was specified in an invalid format.

LifecycleHookLimitExceeded(String)

The limit for lifecycle hooks was exceeded.

TagSetListLimitExceeded(String)

The number of tag groups included in the tag set list exceeded the maximum allowed limit of 3.

TriggerTargetsLimitExceeded(String)

The maximum allowed number of triggers was exceeded.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl UpdateDeploymentGroupError

pub fn from_body(body: &str) -> UpdateDeploymentGroupError