Struct rusoto_codedeploy::BlueGreenDeploymentConfiguration

pub struct BlueGreenDeploymentConfiguration {
    pub deployment_ready_option: Option<DeploymentReadyOption>,
    pub green_fleet_provisioning_option: Option<GreenFleetProvisioningOption>,
    pub terminate_blue_instances_on_deployment_success: Option<BlueInstanceTerminationOption>,
}

Information about blue/green deployment options for a deployment group.

Fields
deployment_ready_option: Option<DeploymentReadyOption>

Information about the action to take when newly provisioned instances are ready to receive traffic in a blue/green deployment.

green_fleet_provisioning_option: Option<GreenFleetProvisioningOption>

Information about how instances are provisioned for a replacement environment in a blue/green deployment.

terminate_blue_instances_on_deployment_success: Option<BlueInstanceTerminationOption>

Information about whether to terminate instances in the original fleet during a blue/green deployment.