Struct rusoto_codedeploy::CodeDeployClient

pub struct CodeDeployClient { /* fields omitted */ }

A client for the CodeDeploy API.

Methods

impl CodeDeployClient

pub fn new(region: Region) -> CodeDeployClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> CodeDeployClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,