Struct rusoto_codedeploy::CreateDeploymentInput

pub struct CreateDeploymentInput {
    pub application_name: String,
    pub auto_rollback_configuration: Option<AutoRollbackConfiguration>,
    pub deployment_config_name: Option<String>,
    pub deployment_group_name: Option<String>,
    pub description: Option<String>,
    pub file_exists_behavior: Option<String>,
    pub ignore_application_stop_failures: Option<bool>,
    pub revision: Option<RevisionLocation>,
    pub target_instances: Option<TargetInstances>,
    pub update_outdated_instances_only: Option<bool>,
}

Represents the input of a CreateDeployment operation.

Fields
application_name: String

The name of an AWS CodeDeploy application associated with the applicable IAM user or AWS account.

auto_rollback_configuration: Option<AutoRollbackConfiguration>

Configuration information for an automatic rollback that is added when a deployment is created.

deployment_config_name: Option<String>

The name of a deployment configuration associated with the applicable IAM user or AWS account.

If not specified, the value configured in the deployment group will be used as the default. If the deployment group does not have a deployment configuration associated with it, then CodeDeployDefault.OneAtATime will be used by default.

deployment_group_name: Option<String>

The name of the deployment group.

description: Option<String>

A comment about the deployment.

file_exists_behavior: Option<String>

Information about how AWS CodeDeploy handles files that already exist in a deployment target location but weren't part of the previous successful deployment.

The fileExistsBehavior parameter takes any of the following values:

DISALLOW: The deployment fails. This is also the default behavior if no option is specified.

OVERWRITE: The version of the file from the application revision currently being deployed replaces the version already on the instance.

RETAIN: The version of the file already on the instance is kept and used as part of the new deployment.

ignore_application_stop_failures: Option<bool>

If set to true, then if the deployment causes the ApplicationStop deployment lifecycle event to an instance to fail, the deployment to that instance will not be considered to have failed at that point and will continue on to the BeforeInstall deployment lifecycle event.

If set to false or not specified, then if the deployment causes the ApplicationStop deployment lifecycle event to fail to an instance, the deployment to that instance will stop, and the deployment to that instance will be considered to have failed.

revision: Option<RevisionLocation>

The type and location of the revision to deploy.

target_instances: Option<TargetInstances>

Information about the instances that will belong to the replacement environment in a blue/green deployment.

update_outdated_instances_only: Option<bool>

Indicates whether to deploy to all instances or only to instances that are not running the latest application revision.