Struct rusoto_codedeploy::DeploymentGroupInfo

pub struct DeploymentGroupInfo {
    pub alarm_configuration: Option<AlarmConfiguration>,
    pub application_name: Option<String>,
    pub auto_rollback_configuration: Option<AutoRollbackConfiguration>,
    pub auto_scaling_groups: Option<Vec<AutoScalingGroup>>,
    pub blue_green_deployment_configuration: Option<BlueGreenDeploymentConfiguration>,
    pub deployment_config_name: Option<String>,
    pub deployment_group_id: Option<String>,
    pub deployment_group_name: Option<String>,
    pub deployment_style: Option<DeploymentStyle>,
    pub ec_2_tag_filters: Option<Vec<EC2TagFilter>>,
    pub last_attempted_deployment: Option<LastDeploymentInfo>,
    pub last_successful_deployment: Option<LastDeploymentInfo>,
    pub load_balancer_info: Option<LoadBalancerInfo>,
    pub on_premises_instance_tag_filters: Option<Vec<TagFilter>>,
    pub service_role_arn: Option<String>,
    pub target_revision: Option<RevisionLocation>,
    pub trigger_configurations: Option<Vec<TriggerConfig>>,
}

Information about a deployment group.

Fields
alarm_configuration: Option<AlarmConfiguration>

A list of alarms associated with the deployment group.

application_name: Option<String>

The application name.

auto_rollback_configuration: Option<AutoRollbackConfiguration>

Information about the automatic rollback configuration associated with the deployment group.

auto_scaling_groups: Option<Vec<AutoScalingGroup>>

A list of associated Auto Scaling groups.

blue_green_deployment_configuration: Option<BlueGreenDeploymentConfiguration>

Information about blue/green deployment options for a deployment group.

deployment_config_name: Option<String>

The deployment configuration name.

deployment_group_id: Option<String>

The deployment group ID.

deployment_group_name: Option<String>

The deployment group name.

deployment_style: Option<DeploymentStyle>

Information about the type of deployment, either in-place or blue/green, you want to run and whether to route deployment traffic behind a load balancer.

ec_2_tag_filters: Option<Vec<EC2TagFilter>>

The Amazon EC2 tags on which to filter.

last_attempted_deployment: Option<LastDeploymentInfo>

Information about the most recent attempted deployment to the deployment group.

last_successful_deployment: Option<LastDeploymentInfo>

Information about the most recent successful deployment to the deployment group.

load_balancer_info: Option<LoadBalancerInfo>

Information about the load balancer to use in a deployment.

on_premises_instance_tag_filters: Option<Vec<TagFilter>>

The on-premises instance tags on which to filter.

service_role_arn: Option<String>

A service role ARN.

target_revision: Option<RevisionLocation>

Information about the deployment group's target revision, including type and location.

trigger_configurations: Option<Vec<TriggerConfig>>

Information about triggers associated with the deployment group.