Struct rusoto_codedeploy::DeploymentInfo

pub struct DeploymentInfo {
    pub additional_deployment_status_info: Option<String>,
    pub application_name: Option<String>,
    pub auto_rollback_configuration: Option<AutoRollbackConfiguration>,
    pub blue_green_deployment_configuration: Option<BlueGreenDeploymentConfiguration>,
    pub complete_time: Option<f64>,
    pub create_time: Option<f64>,
    pub creator: Option<String>,
    pub deployment_config_name: Option<String>,
    pub deployment_group_name: Option<String>,
    pub deployment_id: Option<String>,
    pub deployment_overview: Option<DeploymentOverview>,
    pub deployment_style: Option<DeploymentStyle>,
    pub description: Option<String>,
    pub error_information: Option<ErrorInformation>,
    pub file_exists_behavior: Option<String>,
    pub ignore_application_stop_failures: Option<bool>,
    pub instance_termination_wait_time_started: Option<bool>,
    pub load_balancer_info: Option<LoadBalancerInfo>,
    pub previous_revision: Option<RevisionLocation>,
    pub revision: Option<RevisionLocation>,
    pub rollback_info: Option<RollbackInfo>,
    pub start_time: Option<f64>,
    pub status: Option<String>,
    pub target_instances: Option<TargetInstances>,
    pub update_outdated_instances_only: Option<bool>,
}

Information about a deployment.

Fields
additional_deployment_status_info: Option<String>

Provides information about the results of a deployment, such as whether instances in the original environment in a blue/green deployment were not terminated.

application_name: Option<String>

The application name.

auto_rollback_configuration: Option<AutoRollbackConfiguration>

Information about the automatic rollback configuration associated with the deployment.

blue_green_deployment_configuration: Option<BlueGreenDeploymentConfiguration>

Information about blue/green deployment options for this deployment.

complete_time: Option<f64>

A timestamp indicating when the deployment was complete.

create_time: Option<f64>

A timestamp indicating when the deployment was created.

creator: Option<String>

The means by which the deployment was created:

user: A user created the deployment.

autoscaling: Auto Scaling created the deployment.

codeDeployRollback: A rollback process created the deployment.

deployment_config_name: Option<String>

The deployment configuration name.

deployment_group_name: Option<String>

The deployment group name.

deployment_id: Option<String>

The deployment ID.

deployment_overview: Option<DeploymentOverview>

A summary of the deployment status of the instances in the deployment.

deployment_style: Option<DeploymentStyle>

Information about the type of deployment, either in-place or blue/green, you want to run and whether to route deployment traffic behind a load balancer.

description: Option<String>

A comment about the deployment.

error_information: Option<ErrorInformation>

Information about any error associated with this deployment.

file_exists_behavior: Option<String>

Information about how AWS CodeDeploy handles files that already exist in a deployment target location but weren't part of the previous successful deployment.

DISALLOW: The deployment fails. This is also the default behavior if no option is specified.

OVERWRITE: The version of the file from the application revision currently being deployed replaces the version already on the instance.

RETAIN: The version of the file already on the instance is kept and used as part of the new deployment.

ignore_application_stop_failures: Option<bool>

If true, then if the deployment causes the ApplicationStop deployment lifecycle event to an instance to fail, the deployment to that instance will not be considered to have failed at that point and will continue on to the BeforeInstall deployment lifecycle event.

If false or not specified, then if the deployment causes the ApplicationStop deployment lifecycle event to an instance to fail, the deployment to that instance will stop, and the deployment to that instance will be considered to have failed.

instance_termination_wait_time_started: Option<bool>

Indicates whether the wait period set for the termination of instances in the original environment has started. Status is 'false' if the KEEP_ALIVE option is specified; otherwise, 'true' as soon as the termination wait period starts.

load_balancer_info: Option<LoadBalancerInfo>

Information about the load balancer used in the deployment.

previous_revision: Option<RevisionLocation>

Information about the application revision that was deployed to the deployment group before the most recent successful deployment.

revision: Option<RevisionLocation>

Information about the location of stored application artifacts and the service from which to retrieve them.

rollback_info: Option<RollbackInfo>

Information about a deployment rollback.

start_time: Option<f64>

A timestamp indicating when the deployment was deployed to the deployment group.

In some cases, the reported value of the start time may be later than the complete time. This is due to differences in the clock settings of back-end servers that participate in the deployment process.

status: Option<String>

The current state of the deployment as a whole.

target_instances: Option<TargetInstances>

Information about the instances that belong to the replacement environment in a blue/green deployment.

update_outdated_instances_only: Option<bool>

Indicates whether only instances that are not running the latest application revision are to be deployed to.