Struct rusoto_codedeploy::S3Location

pub struct S3Location {
    pub bucket: Option<String>,
    pub bundle_type: Option<String>,
    pub e_tag: Option<String>,
    pub key: Option<String>,
    pub version: Option<String>,
}

Information about the location of application artifacts stored in Amazon S3.

Fields
bucket: Option<String>

The name of the Amazon S3 bucket where the application revision is stored.

bundle_type: Option<String>

The file type of the application revision. Must be one of the following:

tar: A tar archive file.

tgz: A compressed tar archive file.

zip: A zip archive file.

e_tag: Option<String>

The ETag of the Amazon S3 object that represents the bundled artifacts for the application revision.

If the ETag is not specified as an input parameter, ETag validation of the object will be skipped.

key: Option<String>

The name of the Amazon S3 object that represents the bundled artifacts for the application revision.

version: Option<String>

A specific version of the Amazon S3 object that represents the bundled artifacts for the application revision.

If the version is not specified, the system will use the most recent version by default.