Crate rusoto_codedeploy

AWS CodeDeploy
AWS CodeDeploy is a deployment service that automates application deployments to Amazon EC2 instances, on-premises instances running in your own facility, or serverless AWS Lambda functions.

You can deploy a nearly unlimited variety of application content, such as an updated Lambda function, code, web and configuration files, executables, packages, scripts, multimedia files, and so on. AWS CodeDeploy can deploy application content stored in Amazon S3 buckets, GitHub repositories, or Bitbucket repositories. You do not need to make changes to your existing code before you can use AWS CodeDeploy.

AWS CodeDeploy makes it easier for you to rapidly release new features, helps you avoid downtime during application deployment, and handles the complexity of updating your applications, without many of the risks associated with error-prone manual deployments.

AWS CodeDeploy Components

Use the information in this guide to help you work with the following AWS CodeDeploy components:

Application: A name that uniquely identifies the application you want to deploy. AWS CodeDeploy uses this name, which functions as a container, to ensure the correct combination of revision, deployment configuration, and deployment group are referenced during a deployment.

Deployment group: A set of individual instances or CodeDeploy Lambda applications. A Lambda deployment group contains a group of applications. An EC2/On-premises deployment group contains individually tagged instances, Amazon EC2 instances in Auto Scaling groups, or both.

Deployment configuration: A set of deployment rules and deployment success and failure conditions used by AWS CodeDeploy during a deployment.

Deployment: The process and the components used in the process of updating a Lambda function or of installing content on one or more instances.

Application revisions: For an AWS Lambda deployment, this is an AppSpec file that specifies the Lambda function to update and one or more functions to validate deployment lifecycle events. For an EC2/On-premises deployment, this is an archive file containing source content—source code, web pages, executable files, and deployment scripts—along with an AppSpec file. Revisions are stored in Amazon S3 buckets or GitHub repositories. For Amazon S3, a revision is uniquely identified by its Amazon S3 object key and its ETag, version, or both. For GitHub, a revision is uniquely identified by its commit ID.

This guide also contains information to help you get details about the instances in your deployments, to make on-premises instances available for AWS CodeDeploy deployments, and to get details about a Lambda function deployment.

AWS CodeDeploy Information Resources

AWS CodeDeploy User Guide

AWS CodeDeploy API Reference Guide

AWS CLI Reference for AWS CodeDeploy

AWS CodeDeploy Developer Forum

If you're using the service, you're probably looking for CodeDeployClient and CodeDeploy.

Structs

AddTagsToOnPremisesInstancesInput
Represents the input of, and adds tags to, an on-premises instance operation.

Alarm
Information about an alarm.

AlarmConfiguration
Information about alarms associated with the deployment group.

ApplicationInfo
Information about an application.

AutoRollbackConfiguration
Information about a configuration for automatically rolling back to a previous version of an application revision when a deployment doesn't complete successfully.

AutoScalingGroup
Information about an Auto Scaling group.

BatchGetApplicationRevisionsInput
Represents the input of a BatchGetApplicationRevisions operation.

BatchGetApplicationRevisionsOutput
Represents the output of a BatchGetApplicationRevisions operation.

BatchGetApplicationsInput
Represents the input of a BatchGetApplications operation.

BatchGetApplicationsOutput
Represents the output of a BatchGetApplications operation.

BatchGetDeploymentGroupsInput
Represents the input of a BatchGetDeploymentGroups operation.

BatchGetDeploymentGroupsOutput
Represents the output of a BatchGetDeploymentGroups operation.

BatchGetDeploymentInstancesInput
Represents the input of a BatchGetDeploymentInstances operation.

BatchGetDeploymentInstancesOutput
Represents the output of a BatchGetDeploymentInstances operation.

BatchGetDeploymentsInput
Represents the input of a BatchGetDeployments operation.

BatchGetDeploymentsOutput
Represents the output of a BatchGetDeployments operation.

BatchGetOnPremisesInstancesInput
Represents the input of a BatchGetOnPremisesInstances operation.

BatchGetOnPremisesInstancesOutput
Represents the output of a BatchGetOnPremisesInstances operation.

BlueGreenDeploymentConfiguration
Information about blue/green deployment options for a deployment group.

BlueInstanceTerminationOption
Information about whether instances in the original environment are terminated when a blue/green deployment is successful.

CodeDeployClient
A client for the CodeDeploy API.

ContinueDeploymentInput

CreateApplicationInput
Represents the input of a CreateApplication operation.

CreateApplicationOutput
Represents the output of a CreateApplication operation.

CreateDeploymentConfigInput
Represents the input of a CreateDeploymentConfig operation.

CreateDeploymentConfigOutput
Represents the output of a CreateDeploymentConfig operation.

CreateDeploymentGroupInput
Represents the input of a CreateDeploymentGroup operation.

CreateDeploymentGroupOutput
Represents the output of a CreateDeploymentGroup operation.

CreateDeploymentInput
Represents the input of a CreateDeployment operation.

CreateDeploymentOutput
Represents the output of a CreateDeployment operation.

DeleteApplicationInput
Represents the input of a DeleteApplication operation.

DeleteDeploymentConfigInput
Represents the input of a DeleteDeploymentConfig operation.

DeleteDeploymentGroupInput
Represents the input of a DeleteDeploymentGroup operation.

DeleteDeploymentGroupOutput
Represents the output of a DeleteDeploymentGroup operation.

DeleteGitHubAccountTokenInput
Represents the input of a DeleteGitHubAccount operation.

DeleteGitHubAccountTokenOutput
Represents the output of a DeleteGitHubAccountToken operation.

DeploymentConfigInfo
Information about a deployment configuration.

DeploymentGroupInfo
Information about a deployment group.

DeploymentInfo
Information about a deployment.

DeploymentOverview
Information about the deployment status of the instances in the deployment.

DeploymentReadyOption
Information about how traffic is rerouted to instances in a replacement environment in a blue/green deployment.

DeploymentStyle
Information about the type of deployment, either in-place or blue/green, you want to run and whether to route deployment traffic behind a load balancer.

DeregisterOnPremisesInstanceInput
Represents the input of a DeregisterOnPremisesInstance operation.

Diagnostics
Diagnostic information about executable scripts that are part of a deployment.

EC2TagFilter
Information about an EC2 tag filter.

EC2TagSet
Information about groups of EC2 instance tags.

ELBInfo
Information about a load balancer in Elastic Load Balancing to use in a deployment. Instances are registered directly with a load balancer, and traffic is routed to the load balancer.

ErrorInformation
Information about a deployment error.

GenericRevisionInfo
Information about an application revision.

GetApplicationInput
Represents the input of a GetApplication operation.

GetApplicationOutput
Represents the output of a GetApplication operation.

GetApplicationRevisionInput
Represents the input of a GetApplicationRevision operation.

GetApplicationRevisionOutput
Represents the output of a GetApplicationRevision operation.

GetDeploymentConfigInput
Represents the input of a GetDeploymentConfig operation.

GetDeploymentConfigOutput
Represents the output of a GetDeploymentConfig operation.

GetDeploymentGroupInput
Represents the input of a GetDeploymentGroup operation.

GetDeploymentGroupOutput
Represents the output of a GetDeploymentGroup operation.

GetDeploymentInput
Represents the input of a GetDeployment operation.

GetDeploymentInstanceInput
Represents the input of a GetDeploymentInstance operation.

GetDeploymentInstanceOutput
Represents the output of a GetDeploymentInstance operation.

GetDeploymentOutput
Represents the output of a GetDeployment operation.

GetOnPremisesInstanceInput
Represents the input of a GetOnPremisesInstance operation.

GetOnPremisesInstanceOutput
Represents the output of a GetOnPremisesInstance operation.

GitHubLocation
Information about the location of application artifacts stored in GitHub.

GreenFleetProvisioningOption
Information about the instances that belong to the replacement environment in a blue/green deployment.

InstanceInfo
Information about an on-premises instance.

InstanceSummary
Information about an instance in a deployment.

LastDeploymentInfo
Information about the most recent attempted or successful deployment to a deployment group.

LifecycleEvent
Information about a deployment lifecycle event.

ListApplicationRevisionsInput
Represents the input of a ListApplicationRevisions operation.

ListApplicationRevisionsOutput
Represents the output of a ListApplicationRevisions operation.

ListApplicationsInput
Represents the input of a ListApplications operation.

ListApplicationsOutput
Represents the output of a ListApplications operation.

ListDeploymentConfigsInput
Represents the input of a ListDeploymentConfigs operation.

ListDeploymentConfigsOutput
Represents the output of a ListDeploymentConfigs operation.

ListDeploymentGroupsInput
Represents the input of a ListDeploymentGroups operation.

ListDeploymentGroupsOutput
Represents the output of a ListDeploymentGroups operation.

ListDeploymentInstancesInput
Represents the input of a ListDeploymentInstances operation.

ListDeploymentInstancesOutput
Represents the output of a ListDeploymentInstances operation.

ListDeploymentsInput
Represents the input of a ListDeployments operation.

ListDeploymentsOutput
Represents the output of a ListDeployments operation.

ListGitHubAccountTokenNamesInput
Represents the input of a ListGitHubAccountTokenNames operation.

ListGitHubAccountTokenNamesOutput
Represents the output of a ListGitHubAccountTokenNames operation.

ListOnPremisesInstancesInput
Represents the input of a ListOnPremisesInstances operation.

ListOnPremisesInstancesOutput
Represents the output of list on-premises instances operation.

LoadBalancerInfo
Information about the Elastic Load Balancing load balancer or target group used in a deployment.

MinimumHealthyHosts
Information about minimum healthy instance.

OnPremisesTagSet
Information about groups of on-premises instance tags.

PutLifecycleEventHookExecutionStatusInput

PutLifecycleEventHookExecutionStatusOutput

RawString
A revision for an AWS Lambda deployment that is a YAML-formatted or JSON-formatted string. For AWS Lambda deployments, the revision is the same as the AppSpec file.

RegisterApplicationRevisionInput
Represents the input of a RegisterApplicationRevision operation.

RegisterOnPremisesInstanceInput
Represents the input of the register on-premises instance operation.

RemoveTagsFromOnPremisesInstancesInput
Represents the input of a RemoveTagsFromOnPremisesInstances operation.

RevisionInfo
Information about an application revision.

RevisionLocation
Information about the location of an application revision.

RollbackInfo
Information about a deployment rollback.

S3Location
Information about the location of application artifacts stored in Amazon S3.

SkipWaitTimeForInstanceTerminationInput

StopDeploymentInput
Represents the input of a StopDeployment operation.

StopDeploymentOutput
Represents the output of a StopDeployment operation.

Tag
Information about a tag.

TagFilter
Information about an on-premises instance tag filter.

TargetGroupInfo
Information about a target group in Elastic Load Balancing to use in a deployment. Instances are registered as targets in a target group, and traffic is routed to the target group.

TargetInstances
Information about the instances to be used in the replacement environment in a blue/green deployment.

TimeBasedCanary
A configuration that shifts traffic from one version of a Lambda function to another in two increments. The original and target Lambda function versions are specified in the deployment's AppSpec file.

TimeBasedLinear
A configuration that shifts traffic from one version of a Lambda function to another in equal increments, with an equal number of minutes between each increment. The original and target Lambda function versions are specified in the deployment's AppSpec file.

TimeRange
Information about a time range.

TrafficRoutingConfig
The configuration that specifies how traffic is shifted from one version of a Lambda function to another version during an AWS Lambda deployment.

TriggerConfig
Information about notification triggers for the deployment group.

UpdateApplicationInput
Represents the input of an UpdateApplication operation.

UpdateDeploymentGroupInput
Represents the input of an UpdateDeploymentGroup operation.

UpdateDeploymentGroupOutput
Represents the output of an UpdateDeploymentGroup operation.

Enums

AddTagsToOnPremisesInstancesError
Errors returned by AddTagsToOnPremisesInstances

BatchGetApplicationRevisionsError
Errors returned by BatchGetApplicationRevisions

BatchGetApplicationsError
Errors returned by BatchGetApplications

BatchGetDeploymentGroupsError
Errors returned by BatchGetDeploymentGroups

BatchGetDeploymentInstancesError
Errors returned by BatchGetDeploymentInstances

BatchGetDeploymentsError
Errors returned by BatchGetDeployments

BatchGetOnPremisesInstancesError
Errors returned by BatchGetOnPremisesInstances

ContinueDeploymentError
Errors returned by ContinueDeployment

CreateApplicationError
Errors returned by CreateApplication

CreateDeploymentConfigError
Errors returned by CreateDeploymentConfig

CreateDeploymentError
Errors returned by CreateDeployment

CreateDeploymentGroupError
Errors returned by CreateDeploymentGroup

DeleteApplicationError
Errors returned by DeleteApplication

DeleteDeploymentConfigError
Errors returned by DeleteDeploymentConfig

DeleteDeploymentGroupError
Errors returned by DeleteDeploymentGroup

DeleteGitHubAccountTokenError
Errors returned by DeleteGitHubAccountToken

DeregisterOnPremisesInstanceError
Errors returned by DeregisterOnPremisesInstance

GetApplicationError
Errors returned by GetApplication

GetApplicationRevisionError
Errors returned by GetApplicationRevision

GetDeploymentConfigError
Errors returned by GetDeploymentConfig

GetDeploymentError
Errors returned by GetDeployment

GetDeploymentGroupError
Errors returned by GetDeploymentGroup

GetDeploymentInstanceError
Errors returned by GetDeploymentInstance

GetOnPremisesInstanceError
Errors returned by GetOnPremisesInstance

ListApplicationRevisionsError
Errors returned by ListApplicationRevisions

ListApplicationsError
Errors returned by ListApplications

ListDeploymentConfigsError
Errors returned by ListDeploymentConfigs

ListDeploymentGroupsError
Errors returned by ListDeploymentGroups

ListDeploymentInstancesError
Errors returned by ListDeploymentInstances

ListDeploymentsError
Errors returned by ListDeployments

ListGitHubAccountTokenNamesError
Errors returned by ListGitHubAccountTokenNames

ListOnPremisesInstancesError
Errors returned by ListOnPremisesInstances

PutLifecycleEventHookExecutionStatusError
Errors returned by PutLifecycleEventHookExecutionStatus

RegisterApplicationRevisionError
Errors returned by RegisterApplicationRevision

RegisterOnPremisesInstanceError
Errors returned by RegisterOnPremisesInstance

RemoveTagsFromOnPremisesInstancesError
Errors returned by RemoveTagsFromOnPremisesInstances

SkipWaitTimeForInstanceTerminationError
Errors returned by SkipWaitTimeForInstanceTermination

StopDeploymentError
Errors returned by StopDeployment

UpdateApplicationError
Errors returned by UpdateApplication

UpdateDeploymentGroupError
Errors returned by UpdateDeploymentGroup

Traits

CodeDeploy
Trait representing the capabilities of the CodeDeploy API. CodeDeploy clients implement this trait.