Trait rusoto_codedeploy::CodeDeploy

pub trait CodeDeploy {
    fn add_tags_to_on_premises_instances(
        &self, 
        input: AddTagsToOnPremisesInstancesInput
    ) -> RusotoFuture<(), AddTagsToOnPremisesInstancesError>;

    fn batch_get_application_revisions(
        &self, 
        input: BatchGetApplicationRevisionsInput
    ) -> RusotoFuture<BatchGetApplicationRevisionsOutput, BatchGetApplicationRevisionsError>;

    fn batch_get_applications(
        &self, 
        input: BatchGetApplicationsInput
    ) -> RusotoFuture<BatchGetApplicationsOutput, BatchGetApplicationsError>;

    fn batch_get_deployment_groups(
        &self, 
        input: BatchGetDeploymentGroupsInput
    ) -> RusotoFuture<BatchGetDeploymentGroupsOutput, BatchGetDeploymentGroupsError>;

    fn batch_get_deployment_instances(
        &self, 
        input: BatchGetDeploymentInstancesInput
    ) -> RusotoFuture<BatchGetDeploymentInstancesOutput, BatchGetDeploymentInstancesError>;

    fn batch_get_deployments(
        &self, 
        input: BatchGetDeploymentsInput
    ) -> RusotoFuture<BatchGetDeploymentsOutput, BatchGetDeploymentsError>;

    fn batch_get_on_premises_instances(
        &self, 
        input: BatchGetOnPremisesInstancesInput
    ) -> RusotoFuture<BatchGetOnPremisesInstancesOutput, BatchGetOnPremisesInstancesError>;

    fn continue_deployment(
        &self, 
        input: ContinueDeploymentInput
    ) -> RusotoFuture<(), ContinueDeploymentError>;

    fn create_application(
        &self, 
        input: CreateApplicationInput
    ) -> RusotoFuture<CreateApplicationOutput, CreateApplicationError>;

    fn create_deployment(
        &self, 
        input: CreateDeploymentInput
    ) -> RusotoFuture<CreateDeploymentOutput, CreateDeploymentError>;

    fn create_deployment_config(
        &self, 
        input: CreateDeploymentConfigInput
    ) -> RusotoFuture<CreateDeploymentConfigOutput, CreateDeploymentConfigError>;

    fn create_deployment_group(
        &self, 
        input: CreateDeploymentGroupInput
    ) -> RusotoFuture<CreateDeploymentGroupOutput, CreateDeploymentGroupError>;

    fn delete_application(
        &self, 
        input: DeleteApplicationInput
    ) -> RusotoFuture<(), DeleteApplicationError>;

    fn delete_deployment_config(
        &self, 
        input: DeleteDeploymentConfigInput
    ) -> RusotoFuture<(), DeleteDeploymentConfigError>;

    fn delete_deployment_group(
        &self, 
        input: DeleteDeploymentGroupInput
    ) -> RusotoFuture<DeleteDeploymentGroupOutput, DeleteDeploymentGroupError>;

    fn delete_git_hub_account_token(
        &self, 
        input: DeleteGitHubAccountTokenInput
    ) -> RusotoFuture<DeleteGitHubAccountTokenOutput, DeleteGitHubAccountTokenError>;

    fn deregister_on_premises_instance(
        &self, 
        input: DeregisterOnPremisesInstanceInput
    ) -> RusotoFuture<(), DeregisterOnPremisesInstanceError>;

    fn get_application(
        &self, 
        input: GetApplicationInput
    ) -> RusotoFuture<GetApplicationOutput, GetApplicationError>;

    fn get_application_revision(
        &self, 
        input: GetApplicationRevisionInput
    ) -> RusotoFuture<GetApplicationRevisionOutput, GetApplicationRevisionError>;

    fn get_deployment(
        &self, 
        input: GetDeploymentInput
    ) -> RusotoFuture<GetDeploymentOutput, GetDeploymentError>;

    fn get_deployment_config(
        &self, 
        input: GetDeploymentConfigInput
    ) -> RusotoFuture<GetDeploymentConfigOutput, GetDeploymentConfigError>;

    fn get_deployment_group(
        &self, 
        input: GetDeploymentGroupInput
    ) -> RusotoFuture<GetDeploymentGroupOutput, GetDeploymentGroupError>;

    fn get_deployment_instance(
        &self, 
        input: GetDeploymentInstanceInput
    ) -> RusotoFuture<GetDeploymentInstanceOutput, GetDeploymentInstanceError>;

    fn get_on_premises_instance(
        &self, 
        input: GetOnPremisesInstanceInput
    ) -> RusotoFuture<GetOnPremisesInstanceOutput, GetOnPremisesInstanceError>;

    fn list_application_revisions(
        &self, 
        input: ListApplicationRevisionsInput
    ) -> RusotoFuture<ListApplicationRevisionsOutput, ListApplicationRevisionsError>;

    fn list_applications(
        &self, 
        input: ListApplicationsInput
    ) -> RusotoFuture<ListApplicationsOutput, ListApplicationsError>;

    fn list_deployment_configs(
        &self, 
        input: ListDeploymentConfigsInput
    ) -> RusotoFuture<ListDeploymentConfigsOutput, ListDeploymentConfigsError>;

    fn list_deployment_groups(
        &self, 
        input: ListDeploymentGroupsInput
    ) -> RusotoFuture<ListDeploymentGroupsOutput, ListDeploymentGroupsError>;

    fn list_deployment_instances(
        &self, 
        input: ListDeploymentInstancesInput
    ) -> RusotoFuture<ListDeploymentInstancesOutput, ListDeploymentInstancesError>;

    fn list_deployments(
        &self, 
        input: ListDeploymentsInput
    ) -> RusotoFuture<ListDeploymentsOutput, ListDeploymentsError>;

    fn list_git_hub_account_token_names(
        &self, 
        input: ListGitHubAccountTokenNamesInput
    ) -> RusotoFuture<ListGitHubAccountTokenNamesOutput, ListGitHubAccountTokenNamesError>;

    fn list_on_premises_instances(
        &self, 
        input: ListOnPremisesInstancesInput
    ) -> RusotoFuture<ListOnPremisesInstancesOutput, ListOnPremisesInstancesError>;

    fn put_lifecycle_event_hook_execution_status(
        &self, 
        input: PutLifecycleEventHookExecutionStatusInput
    ) -> RusotoFuture<PutLifecycleEventHookExecutionStatusOutput, PutLifecycleEventHookExecutionStatusError>;

    fn register_application_revision(
        &self, 
        input: RegisterApplicationRevisionInput
    ) -> RusotoFuture<(), RegisterApplicationRevisionError>;

    fn register_on_premises_instance(
        &self, 
        input: RegisterOnPremisesInstanceInput
    ) -> RusotoFuture<(), RegisterOnPremisesInstanceError>;

    fn remove_tags_from_on_premises_instances(
        &self, 
        input: RemoveTagsFromOnPremisesInstancesInput
    ) -> RusotoFuture<(), RemoveTagsFromOnPremisesInstancesError>;

    fn skip_wait_time_for_instance_termination(
        &self, 
        input: SkipWaitTimeForInstanceTerminationInput
    ) -> RusotoFuture<(), SkipWaitTimeForInstanceTerminationError>;

    fn stop_deployment(
        &self, 
        input: StopDeploymentInput
    ) -> RusotoFuture<StopDeploymentOutput, StopDeploymentError>;

    fn update_application(
        &self, 
        input: UpdateApplicationInput
    ) -> RusotoFuture<(), UpdateApplicationError>;

    fn update_deployment_group(
        &self, 
        input: UpdateDeploymentGroupInput
    ) -> RusotoFuture<UpdateDeploymentGroupOutput, UpdateDeploymentGroupError>;
}

Trait representing the capabilities of the CodeDeploy API. CodeDeploy clients implement this trait.

Required Methods

fn add_tags_to_on_premises_instances(
&self,
input: AddTagsToOnPremisesInstancesInput
) -> RusotoFuture<(), AddTagsToOnPremisesInstancesError>

Adds tags to on-premises instances.

fn batch_get_application_revisions(
&self,
input: BatchGetApplicationRevisionsInput
) -> RusotoFuture<BatchGetApplicationRevisionsOutput, BatchGetApplicationRevisionsError>

Gets information about one or more application revisions.

fn batch_get_applications(
&self,
input: BatchGetApplicationsInput
) -> RusotoFuture<BatchGetApplicationsOutput, BatchGetApplicationsError>

Gets information about one or more applications.

fn batch_get_deployment_groups(
&self,
input: BatchGetDeploymentGroupsInput
) -> RusotoFuture<BatchGetDeploymentGroupsOutput, BatchGetDeploymentGroupsError>

Gets information about one or more deployment groups.

fn batch_get_deployment_instances(
&self,
input: BatchGetDeploymentInstancesInput
) -> RusotoFuture<BatchGetDeploymentInstancesOutput, BatchGetDeploymentInstancesError>

Gets information about one or more instance that are part of a deployment group.

fn batch_get_deployments(
&self,
input: BatchGetDeploymentsInput
) -> RusotoFuture<BatchGetDeploymentsOutput, BatchGetDeploymentsError>

Gets information about one or more deployments.

fn batch_get_on_premises_instances(
&self,
input: BatchGetOnPremisesInstancesInput
) -> RusotoFuture<BatchGetOnPremisesInstancesOutput, BatchGetOnPremisesInstancesError>

Gets information about one or more on-premises instances.

fn continue_deployment(
&self,
input: ContinueDeploymentInput
) -> RusotoFuture<(), ContinueDeploymentError>

For a blue/green deployment, starts the process of rerouting traffic from instances in the original environment to instances in the replacement environment without waiting for a specified wait time to elapse. (Traffic rerouting, which is achieved by registering instances in the replacement environment with the load balancer, can start as soon as all instances have a status of Ready.)

fn create_application(
&self,
input: CreateApplicationInput
) -> RusotoFuture<CreateApplicationOutput, CreateApplicationError>

Creates an application.

fn create_deployment(
&self,
input: CreateDeploymentInput
) -> RusotoFuture<CreateDeploymentOutput, CreateDeploymentError>

Deploys an application revision through the specified deployment group.

fn create_deployment_config(
&self,
input: CreateDeploymentConfigInput
) -> RusotoFuture<CreateDeploymentConfigOutput, CreateDeploymentConfigError>

Creates a deployment configuration.

fn create_deployment_group(
&self,
input: CreateDeploymentGroupInput
) -> RusotoFuture<CreateDeploymentGroupOutput, CreateDeploymentGroupError>

Creates a deployment group to which application revisions will be deployed.

fn delete_application(
&self,
input: DeleteApplicationInput
) -> RusotoFuture<(), DeleteApplicationError>

Deletes an application.

fn delete_deployment_config(
&self,
input: DeleteDeploymentConfigInput
) -> RusotoFuture<(), DeleteDeploymentConfigError>

Deletes a deployment configuration.

A deployment configuration cannot be deleted if it is currently in use. Predefined configurations cannot be deleted.

fn delete_deployment_group(
&self,
input: DeleteDeploymentGroupInput
) -> RusotoFuture<DeleteDeploymentGroupOutput, DeleteDeploymentGroupError>

Deletes a deployment group.

fn delete_git_hub_account_token(
&self,
input: DeleteGitHubAccountTokenInput
) -> RusotoFuture<DeleteGitHubAccountTokenOutput, DeleteGitHubAccountTokenError>

Deletes a GitHub account connection.

fn deregister_on_premises_instance(
&self,
input: DeregisterOnPremisesInstanceInput
) -> RusotoFuture<(), DeregisterOnPremisesInstanceError>

Deregisters an on-premises instance.

fn get_application(
&self,
input: GetApplicationInput
) -> RusotoFuture<GetApplicationOutput, GetApplicationError>

Gets information about an application.

fn get_application_revision(
&self,
input: GetApplicationRevisionInput
) -> RusotoFuture<GetApplicationRevisionOutput, GetApplicationRevisionError>

Gets information about an application revision.

fn get_deployment(
&self,
input: GetDeploymentInput
) -> RusotoFuture<GetDeploymentOutput, GetDeploymentError>

Gets information about a deployment.

fn get_deployment_config(
&self,
input: GetDeploymentConfigInput
) -> RusotoFuture<GetDeploymentConfigOutput, GetDeploymentConfigError>

Gets information about a deployment configuration.

fn get_deployment_group(
&self,
input: GetDeploymentGroupInput
) -> RusotoFuture<GetDeploymentGroupOutput, GetDeploymentGroupError>

Gets information about a deployment group.

fn get_deployment_instance(
&self,
input: GetDeploymentInstanceInput
) -> RusotoFuture<GetDeploymentInstanceOutput, GetDeploymentInstanceError>

Gets information about an instance as part of a deployment.

fn get_on_premises_instance(
&self,
input: GetOnPremisesInstanceInput
) -> RusotoFuture<GetOnPremisesInstanceOutput, GetOnPremisesInstanceError>

Gets information about an on-premises instance.

fn list_application_revisions(
&self,
input: ListApplicationRevisionsInput
) -> RusotoFuture<ListApplicationRevisionsOutput, ListApplicationRevisionsError>

Lists information about revisions for an application.

fn list_applications(
&self,
input: ListApplicationsInput
) -> RusotoFuture<ListApplicationsOutput, ListApplicationsError>

Lists the applications registered with the applicable IAM user or AWS account.

fn list_deployment_configs(
&self,
input: ListDeploymentConfigsInput
) -> RusotoFuture<ListDeploymentConfigsOutput, ListDeploymentConfigsError>

Lists the deployment configurations with the applicable IAM user or AWS account.

fn list_deployment_groups(
&self,
input: ListDeploymentGroupsInput
) -> RusotoFuture<ListDeploymentGroupsOutput, ListDeploymentGroupsError>

Lists the deployment groups for an application registered with the applicable IAM user or AWS account.

fn list_deployment_instances(
&self,
input: ListDeploymentInstancesInput
) -> RusotoFuture<ListDeploymentInstancesOutput, ListDeploymentInstancesError>

Lists the instance for a deployment associated with the applicable IAM user or AWS account.

fn list_deployments(
&self,
input: ListDeploymentsInput
) -> RusotoFuture<ListDeploymentsOutput, ListDeploymentsError>

Lists the deployments in a deployment group for an application registered with the applicable IAM user or AWS account.

fn list_git_hub_account_token_names(
&self,
input: ListGitHubAccountTokenNamesInput
) -> RusotoFuture<ListGitHubAccountTokenNamesOutput, ListGitHubAccountTokenNamesError>

Lists the names of stored connections to GitHub accounts.

fn list_on_premises_instances(
&self,
input: ListOnPremisesInstancesInput
) -> RusotoFuture<ListOnPremisesInstancesOutput, ListOnPremisesInstancesError>

Gets a list of names for one or more on-premises instances.

Unless otherwise specified, both registered and deregistered on-premises instance names will be listed. To list only registered or deregistered on-premises instance names, use the registration status parameter.

fn put_lifecycle_event_hook_execution_status(
&self,
input: PutLifecycleEventHookExecutionStatusInput
) -> RusotoFuture<PutLifecycleEventHookExecutionStatusOutput, PutLifecycleEventHookExecutionStatusError>

Sets the result of a Lambda validation function. The function validates one or both lifecycle events (BeforeAllowTraffic and AfterAllowTraffic) and returns Succeeded or Failed.

fn register_application_revision(
&self,
input: RegisterApplicationRevisionInput
) -> RusotoFuture<(), RegisterApplicationRevisionError>

Registers with AWS CodeDeploy a revision for the specified application.

fn register_on_premises_instance(
&self,
input: RegisterOnPremisesInstanceInput
) -> RusotoFuture<(), RegisterOnPremisesInstanceError>

Registers an on-premises instance.

Only one IAM ARN (an IAM session ARN or IAM user ARN) is supported in the request. You cannot use both.

fn remove_tags_from_on_premises_instances(
&self,
input: RemoveTagsFromOnPremisesInstancesInput
) -> RusotoFuture<(), RemoveTagsFromOnPremisesInstancesError>

Removes one or more tags from one or more on-premises instances.

fn skip_wait_time_for_instance_termination(
&self,
input: SkipWaitTimeForInstanceTerminationInput
) -> RusotoFuture<(), SkipWaitTimeForInstanceTerminationError>

In a blue/green deployment, overrides any specified wait time and starts terminating instances immediately after the traffic routing is completed.

fn stop_deployment(
&self,
input: StopDeploymentInput
) -> RusotoFuture<StopDeploymentOutput, StopDeploymentError>

Attempts to stop an ongoing deployment.

fn update_application(
&self,
input: UpdateApplicationInput
) -> RusotoFuture<(), UpdateApplicationError>

Changes the name of an application.

fn update_deployment_group(
&self,
input: UpdateDeploymentGroupInput
) -> RusotoFuture<UpdateDeploymentGroupOutput, UpdateDeploymentGroupError>

Changes information about a deployment group.